+++ pkg/framework/plugins/removepodsviolatinginterpodantiaffinity/pod_antiaffinity.go
@@ -24,11 +24,11 @@ import (
 	"k8s.io/apimachinery/pkg/util/sets"
 	"sigs.k8s.io/descheduler/pkg/descheduler/evictions"
 	podutil "sigs.k8s.io/descheduler/pkg/descheduler/pod"
-	frameworktypes "sigs.k8s.io/descheduler/pkg/framework/types"
 	"sigs.k8s.io/descheduler/pkg/utils"
 
+	frameworktypes "sigs.k8s.io/descheduler/pkg/framework/types"
+
 	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/klog/v2"
 )
 
@@ -86,8 +86,8 @@ func (d *RemovePodsViolatingInterPodAntiAffinity) Deschedule(ctx context.Context
 	}
 
 	podsInANamespace := podutil.GroupByNamespace(pods)
-	podsOnANode := groupByNodeName(pods)
-	nodeMap := createNodeMap(nodes)
+	podsOnANode := podutil.GroupByNodeName(pods)
+	nodeMap := utils.CreateNodeMap(nodes)
 
 loop:
 	for _, node := range nodes {
@@ -97,15 +97,17 @@ loop:
 		podutil.SortPodsBasedOnPriorityLowToHigh(pods)
 		totalPods := len(pods)
 		for i := 0; i < totalPods; i++ {
-			if checkPodsWithAntiAffinityExist(pods[i], podsInANamespace, nodeMap) && d.handle.Evictor().Filter(pods[i]) && d.handle.Evictor().PreEvictionFilter(pods[i]) {
-				if d.handle.Evictor().Evict(ctx, pods[i], evictions.EvictOptions{StrategyName: PluginName}) {
-					// Since the current pod is evicted all other pods which have anti-affinity with this
-					// pod need not be evicted.
-					// Update allPods.
-					podsInANamespace = removePodFromNamespaceMap(pods[i], podsInANamespace)
-					pods = append(pods[:i], pods[i+1:]...)
-					i--
-					totalPods--
+			if utils.CheckPodsWithAntiAffinityExist(pods[i], podsInANamespace, nodeMap) {
+				if d.handle.Evictor().Filter(pods[i]) && d.handle.Evictor().PreEvictionFilter(pods[i]) {
+					if d.handle.Evictor().Evict(ctx, pods[i], evictions.EvictOptions{StrategyName: PluginName}) {
+						// Since the current pod is evicted all other pods which have anti-affinity with this
+						// pod need not be evicted.
+						// Update allPods.
+						podsInANamespace = removePodFromNamespaceMap(pods[i], podsInANamespace)
+						pods = append(pods[:i], pods[i+1:]...)
+						i--
+						totalPods--
+					}
 				}
 			}
 			if d.handle.Evictor().NodeLimitExceeded(node) {
@@ -129,94 +131,4 @@ func removePodFromNamespaceMap(podToRemove *v1.Pod, podMap map[string][]*v1.Pod)
 		}
 	}
 	return podMap
-}
-
-func groupByNodeName(pods []*v1.Pod) map[string][]*v1.Pod {
-	m := make(map[string][]*v1.Pod)
-	for i := 0; i < len(pods); i++ {
-		pod := pods[i]
-		m[pod.Spec.NodeName] = append(m[pod.Spec.NodeName], pod)
-	}
-	return m
-}
-
-func createNodeMap(nodes []*v1.Node) map[string]*v1.Node {
-	m := make(map[string]*v1.Node, len(nodes))
-	for _, node := range nodes {
-		m[node.GetName()] = node
-	}
-	return m
-}
-
-// checkPodsWithAntiAffinityExist checks if there are other pods on the node that the current pod cannot tolerate.
-func checkPodsWithAntiAffinityExist(pod *v1.Pod, pods map[string][]*v1.Pod, nodeMap map[string]*v1.Node) bool {
-	affinity := pod.Spec.Affinity
-	if affinity != nil && affinity.PodAntiAffinity != nil {
-		for _, term := range getPodAntiAffinityTerms(affinity.PodAntiAffinity) {
-			namespaces := utils.GetNamespacesFromPodAffinityTerm(pod, &term)
-			selector, err := metav1.LabelSelectorAsSelector(term.LabelSelector)
-			if err != nil {
-				klog.ErrorS(err, "Unable to convert LabelSelector into Selector")
-				return false
-			}
-			for namespace := range namespaces {
-				for _, existingPod := range pods[namespace] {
-					if existingPod.Name != pod.Name && utils.PodMatchesTermsNamespaceAndSelector(existingPod, namespaces, selector) {
-						node, ok := nodeMap[pod.Spec.NodeName]
-						if !ok {
-							continue
-						}
-						nodeHavingExistingPod, ok := nodeMap[existingPod.Spec.NodeName]
-						if !ok {
-							continue
-						}
-						if hasSameLabelValue(node, nodeHavingExistingPod, term.TopologyKey) {
-							klog.V(1).InfoS("Found Pods violating PodAntiAffinity", "pod to evicted", klog.KObj(pod))
-							return true
-						}
-					}
-				}
-			}
-		}
-	}
-	return false
-}
-
-// hasSameLabelValue checks if the pods are in the same topology zone.
-func hasSameLabelValue(node1, node2 *v1.Node, key string) bool {
-	if node1.Name == node2.Name {
-		return true
-	}
-
-	// no match if node has empty labels
-	node1Labels := node1.Labels
-	if node1Labels == nil {
-		return false
-	}
-	node2Labels := node2.Labels
-	if node2Labels == nil {
-		return false
-	}
-
-	// no match if node has no topology zone label with given key
-	value1, ok := node1Labels[key]
-	if !ok {
-		return false
-	}
-	value2, ok := node2Labels[key]
-	if !ok {
-		return false
-	}
-
-	return value1 == value2
-}
-
-// getPodAntiAffinityTerms gets the antiaffinity terms for the given pod.
-func getPodAntiAffinityTerms(podAntiAffinity *v1.PodAntiAffinity) (terms []v1.PodAffinityTerm) {
-	if podAntiAffinity != nil {
-		if len(podAntiAffinity.RequiredDuringSchedulingIgnoredDuringExecution) != 0 {
-			terms = podAntiAffinity.RequiredDuringSchedulingIgnoredDuringExecution
-		}
-	}
-	return terms
 }
