+++ pkg/utils/priority.go
@@ -14,10 +14,10 @@ import (
 
 const SystemCriticalPriority = 2 * int32(1000000000)
 
-// GetNamespacesFromPodAffinityTerm returns a set of names
+// getNamespacesFromPodAffinityTerm returns a set of names
 // according to the namespaces indicated in podAffinityTerm.
 // If namespaces is empty it considers the given pod's namespace.
-func GetNamespacesFromPodAffinityTerm(pod *v1.Pod, podAffinityTerm *v1.PodAffinityTerm) sets.Set[string] {
+func getNamespacesFromPodAffinityTerm(pod *v1.Pod, podAffinityTerm *v1.PodAffinityTerm) sets.Set[string] {
 	names := sets.New[string]()
 	if len(podAffinityTerm.Namespaces) == 0 {
 		names.Insert(pod.Namespace)
@@ -27,9 +27,9 @@ func GetNamespacesFromPodAffinityTerm(pod *v1.Pod, podAffinityTerm *v1.PodAffini
 	return names
 }
 
-// PodMatchesTermsNamespaceAndSelector returns true if the given <pod>
+// podMatchesTermsNamespaceAndSelector returns true if the given <pod>
 // matches the namespace and selector defined by <affinityPod>`s <term>.
-func PodMatchesTermsNamespaceAndSelector(pod *v1.Pod, namespaces sets.Set[string], selector labels.Selector) bool {
+func podMatchesTermsNamespaceAndSelector(pod *v1.Pod, namespaces sets.Set[string], selector labels.Selector) bool {
 	if !namespaces.Has(pod.Namespace) {
 		return false
 	}
